REF: dual From trait
